repair long writes for VXI-11

+++ SW/src/AR488_GPIBbus.h
@@ -215,7 +215,7 @@ public:
   enum gpibHandshakeState readByte(uint8_t *db, bool readWithEoi, bool *eoi);
   enum gpibHandshakeState writeByte(uint8_t db, bool isLastByte);
   enum receiveState receiveData(Stream &dataStream, bool detectEoi, bool detectEndByte, uint8_t endByte, size_t maxSize = 0);
-  void sendData(const char *data, uint8_t dsize, bool isLastPacket = true);
+  void sendData(const char *data, uint16_t dsize, bool isLastPacket = true);
 //  void clearDataBus();
   void setControlVal(uint8_t value);
   void setDataVal(uint8_t value);
